+++ backend/src/eeg/mod.rs
@@ -12,53 +12,53 @@ use crate::error::TimeSeriesProcessingError;
 /// Retrieves the power density of the given time series based on the given timestep (ms)
 /// and total time elapsed by the end of the series (ms), returns tuple of power spectrum 
 /// and associated frequency respectively
-pub fn get_power_density(x: Vec<f64>, dt: f64, total_time: f64) -> (Array1<f64>, Array1<f64>) {
-    let x_mean = x.iter().sum::<f64>() / x.len() as f64;
+pub fn get_power_density(x: Vec<f32>, dt: f32, total_time: f32) -> (Array1<f32>, Array1<f32>) {
+    let x_mean = x.iter().sum::<f32>() / x.len() as f32;
 
     let mut planner = FftPlanner::new();
     let fft = planner.plan_fft(x.len(), FftDirection::Forward);
     
-    let mut x_fft: Array1<Complex<f64>> = Array1::zeros(x.len());
+    let mut x_fft: Array1<Complex<f32>> = Array1::zeros(x.len());
     for (i, &x_i) in x.iter().enumerate() {
         x_fft[i] = Complex::new(x_i - x_mean, 0.0);
     }
     
-    let x_fft_slice: &mut [Complex<f64>] = x_fft.as_slice_mut().unwrap();
+    let x_fft_slice: &mut [Complex<f32>] = x_fft.as_slice_mut().unwrap();
     fft.process(x_fft_slice);
 
-    let x_fft_array: Array1<Complex<f64>> = x_fft_slice.to_owned().into();
+    let x_fft_array: Array1<Complex<f32>> = x_fft_slice.to_owned().into();
 
-    let sxx: Array1<Complex<f64>> = x_fft_array.mapv(|val| {
+    let sxx: Array1<Complex<f32>> = x_fft_array.mapv(|val| {
         let conj = val.conj();
-        2.0 * dt.powi(2) / (x.len() as f64 * dt) * (val * conj)
+        2.0 * dt.powi(2) / (x.len() as f32 * dt) * (val * conj)
     });
 
     let sxx_positive = sxx.slice(s![0..(x.len() / 2)]);
     let sxx_positive = sxx_positive.mapv(|val| val.re);
 
-    let df: f64 = 1.0 / total_time as f64;
+    let df: f32 = 1.0 / total_time as f32;
 
-    let fnq: f64 = 1.0 / (2.0 * dt);
+    let fnq: f32 = 1.0 / (2.0 * dt);
 
-    let faxis: Array1<f64> = Array1::range(0.0, fnq, df);
+    let faxis: Array1<f32> = Array1::range(0.0, fnq, df);
 
     return (faxis, sxx_positive.to_owned())
 }
 
-fn find_max(arr: &Array1<f64>) -> Option<&f64> {
+fn find_max(arr: &Array1<f32>) -> Option<&f32> {
     arr.iter().max_by(|a, b| a.total_cmp(b))
 }
 
 /// Compares two power densities spectra using the earth moving distance, 
 /// it assumes the same frequency range for each argument, 
 /// (only compares the second item of [`get_power_density`])
-pub fn power_density_comparison(sxx1: &Array1<f64>, sxx2: &Array1<f64>) -> Result<f64, TimeSeriesProcessingError> {
+pub fn power_density_comparison(sxx1: &Array1<f32>, sxx2: &Array1<f32>) -> Result<f32, TimeSeriesProcessingError> {
     if sxx1.len() != sxx2.len() {
         return Err(TimeSeriesProcessingError::SeriesAreNotSameLength);
     }
 
-    let values = (0..sxx1.len()).map(|x| x as f64)
-        .collect::<Vec<f64>>();
+    let values = (0..sxx1.len()).map(|x| x as f32)
+        .collect::<Vec<f32>>();
 
     let u_values = Array1::from(values.clone());
     let v_values = Array1::from(values);
